# api/openapi.yaml
openapi: 3.1.0
info:
  title: vuhive-cloud Control Plane REST API
  description: |
    Cloud-native orchestration engine for distributed high-throughput load testing workloads on Kubernetes.
    Provides APIs for test suite compilation, runner profiles, scheduled execution (CronJobs),
    ad-hoc execution, start barrier synchronization, and historical telemetry / report query.

    All endpoints support Cross-Origin Resource Sharing (CORS) and respond to HTTP OPTIONS preflight
    requests with HTTP 204 No Content, allowing web browsers and tools like Swagger UI to explore and
    interact with the API across different domains and ports.
  version: 0.0.1
servers:
  - url: http://localhost:8080
    description: Local port-forwarded control plane
  - url: http://localhost:8081
    description: Local Backend-For-Frontend (BFF) service
  - url: http://vuhive-cloud.vuhive-system.svc.cluster.local:8080
    description: In-cluster Kubernetes service
security:
  - BearerAuth: []
  - OAuth2AuthorizationCode:
      - openid
      - profile
      - email
paths:
...
  /api/v1/runs:
    get:
      summary: List and filter test runs
      description: |
        Queries historical test executions with filtering by test suite, execution status,
        originating schedule, and timestamp range. Supports limit/offset pagination and sorts
        results by creation timestamp descending.
      operationId: listRuns
      tags:
        - Runs
      parameters:
        - name: suite_id
          in: query
          description: Filter by parent Test Suite ID
          required: false
          schema:
            type: string
        - name: status
          in: query
          description: Filter by execution status
          required: false
          schema:
            type: string
            enum: [QUEUED, RUNNING, COMPLETED, FAILED, ABORTED]
        - name: schedule_id
          in: query
          description: Filter by recurring Schedule ID
          required: false
          schema:
            type: string
        - name: from
          in: query
          description: Filter test runs created at or after timestamp (RFC 3339)
          required: false
          schema:
            type: string
            format: date-time
            example: "2026-09-01T00:00:00Z"
        - name: to
          in: query
          description: Filter test runs created at or before timestamp (RFC 3339)
          required: false
          schema:
            type: string
            format: date-time
            example: "2026-09-05T23:59:59Z"
        - name: limit
          in: query
          description: Maximum number of runs to return (default 20, max 100)
          required: false
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - name: offset
          in: query
          description: Number of records to skip for pagination
          required: false
          schema:
            type: integer
            default: 0
            minimum: 0
      responses:
        '200':
          description: Paginated list of test executions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunListResponse'
    post:
      summary: Trigger an ad-hoc test run
      description: |
        Dispatches an immediate ad-hoc test run execution. Validates the test suite,
        compiled artifact, and runner profile, creates a TestRun record in QUEUED status,
        and manifests an ephemeral Kubernetes batch/v1 Job in the runner namespace.
      operationId: triggerRun
      tags:
        - Runs
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TriggerRunRequest'
      responses:
        '201':
          description: Test run created and Kubernetes Job dispatched
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunResponse'
        '400':
          description: Invalid request payload or validation failure
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Suite, artifact, configuration, or runner profile not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Resource state conflict (e.g. inactive suite or unready artifact)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/runs/{id}:
    get:
      summary: Retrieve test run details and performance KPIs
      description: |
        Returns complete metadata, execution timestamps, duration in milliseconds, exit code,
        SLA compliance, and indexed summary metrics directly from PostgreSQL without reading S3.
      operationId: getRun
      tags:
        - Runs
      parameters:
        - name: id
          in: path
          required: true
          description: Unique identifier of the test run
          schema:
            type: string
      responses:
        '200':
          description: Test execution details and performance KPIs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunResponse'
        '404':
          description: Test run not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/runs/{id}/report:
    get:
      summary: Download full summary.json report
      description: |
        Fetches and returns the full raw summary.json artifact stored in object storage (or indexed in PostgreSQL).
        If presign=true, returns a secure presigned S3 download URL.
      operationId: getRunReport
      tags:
        - Runs
      parameters:
        - name: id
          in: path
          required: true
          description: Unique identifier of the test run
          schema:
            type: string
        - name: presign
          in: query
          description: When true, returns a presigned S3 download URL instead of streaming content
          required: false
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Raw summary.json content or presigned URL response
          content:
            application/json:
              schema:
                oneOf:
                  - type: object
                    description: Raw summary JSON report
                  - $ref: '#/components/schemas/PresignedURLResponse'
        '404':
          description: Test run or report artifact not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Test run is still in-flight (QUEUED or RUNNING); report not yet generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/runs/{id}/logs:
    get:
      summary: Retrieve execution logs (run.log)
      description: |
        Streams or returns the raw container execution log output (run.log) stored in object storage.
        If presign=true, returns a secure presigned S3 download URL.
      operationId: getRunLogs
      tags:
        - Runs
      parameters:
        - name: id
          in: path
          required: true
          description: Unique identifier of the test run
          schema:
            type: string
        - name: presign
          in: query
          description: When true, returns a presigned S3 download URL instead of streaming content
          required: false
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Raw text log stream or presigned URL response
          content:
            text/plain:
              schema:
                type: string
            application/json:
              schema:
                $ref: '#/components/schemas/PresignedURLResponse'
        '404':
          description: Test run or execution log not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Test run is still in-flight; logs not yet finalized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/runs/{id}/abort:
    parameters:
      - name: id
        in: path
        required: true
        description: Unique TestRun UUID to abort
        schema:
          type: string
    post:
      summary: Abort in-flight test run
      description: |
        Stops and cancels an active (QUEUED or RUNNING) test run on demand.
        Deletes the underlying Kubernetes batch/v1 Job and terminates associated runner pods
        with SIGTERM signal propagation, allowing the runner wrapper to flush partial execution
        logs before pod eviction. Transitions the aggregate state to ABORTED and records audit metadata.
        Rejects abort requests on already finalized runs with 409 Conflict.
      operationId: abortRun
      tags:
        - Runs
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AbortRunRequest'
      responses:
        '200':
          description: Test execution successfully aborted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunResponse'
        '400':
          description: Malformed or invalid request payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Test run not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Run is already in a terminal status (COMPLETED, FAILED, or ABORTED)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/runs/{id}/complete:
    post:
      summary: Finalize test run execution and ingest summary report
      description: |
        Internal runner callback endpoint invoked upon test completion to upload summary metrics,
        record exit status, SLA evaluation, and transition the test run lifecycle state.

        The `{id}` path parameter accepts either a domain UUID (for directly-dispatched ad-hoc runs)
        or a Kubernetes Job name (for CronJob-spawned runs). When a UUID lookup returns not-found,
        the control plane automatically retries resolution by `k8s_job_name`.
      operationId: completeRun
      tags:
        - Runs
      parameters:
        - name: id
          in: path
          required: true
          description: Unique identifier of the test run
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompleteRunRequest'
      responses:
        '200':
          description: Test run successfully finalized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunResponse'
        '400':
          description: Malformed payload or validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Test run not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Test run already in terminal state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/runs/complete:
    post:
      summary: Finalize test run execution and ingest summary report (callback without path ID)
      description: |
        Internal runner callback endpoint invoked upon test completion where `run_id` is supplied
        in the request body payload. This is the **primary callback target** configured via
        `API_CALLBACK_URL` in Kubernetes runner workloads.

        The `run_id` field accepts either a domain UUID (ad-hoc runs) or a Kubernetes Job name
        (CronJob-spawned runs, where `VUHIVE_RUN_ID` resolves to
        `metadata.labels['batch.kubernetes.io/job-name']`). The control plane will attempt UUID
        lookup first, then fall back to `k8s_job_name` correlation automatically.
      operationId: completeRunFromBody
      tags:
        - Runs
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompleteRunRequest'
      responses:
        '200':
          description: Test run successfully finalized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunResponse'
        '400':
          description: Malformed payload or validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Test run not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Test run already in terminal state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/runs/{id}/barrier/await:
    post:
      summary: Await start barrier rendezvous
      description: Participates in a distributed start barrier synchronization for multi-pod load tests.
      operationId: awaitBarrier
      tags:
        - Barrier
      parameters:
        - name: id
          in: path
          required: true
          description: Unique identifier of the test run
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BarrierAwaitRequest'
      responses:
        '200':
          description: Barrier rendezvous released
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BarrierResponse'
        '408':
          description: Barrier rendezvous timed out waiting for all participants
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '424':
          description: Barrier rendezvous was aborted by another participant
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/runs/{id}/barrier/abort:
    post:
      summary: Abort start barrier rendezvous
      description: Signals an abort to release all waiting workers due to initialization failure.
      operationId: abortBarrier
      tags:
        - Barrier
      parameters:
        - name: id
          in: path
          required: true
          description: Unique identifier of the test run
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BarrierAbortRequest'
      responses:
        '200':
          description: Barrier rendezvous successfully aborted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BarrierResponse'
  /api/v1/runs/{id}/barrier:
    get:
      summary: Query barrier status
      description: Inspects current participant readiness and target start time for a distributed test run.
      operationId: getBarrier
      tags:
        - Barrier
      parameters:
        - name: id
          in: path
          required: true
          description: Unique identifier of the test run
          schema:
            type: string
      responses:
        '200':
          description: Barrier session status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BarrierResponse'
  /api/v1/suites/{id}/builds:
    post:
      summary: Upload Go test source and trigger build Job
      description: |
        Accepts a multipart/form-data payload containing the Go test source archive
        (`.tar.gz`) and triggers an ephemeral Kubernetes `batch/v1` Job to compile it
        for the requested target platform. The response is returned immediately; the
        build runs asynchronously.

        **Pre-Build AST Static Analysis & Framework Enforcement**:
        Uploaded archives undergo strict automated static verification before compilation:
        - **Direct Dependency Check**: `go.mod` must declare `github.com/morphy76/vuhive` as a direct dependency.
        - **Inverted Control Driver**: Source code must declare `package scenario` (defining a valid `vuhive.Scenario` entrypoint such as `NewScenario()`, `Scenario()`, `InitScenario()`, or `Register(*vuhive.Engine)`). Defining `package main` or `func main()` is prohibited; the platform injects a trusted `main.go` execution driver.
        - **Package Import Blocklist**: Dangerous packages (`os/exec`, `syscall`, `unsafe`, `plugin`, `runtime/cgo`, `golang.org/x/sys`) are blocked by default. Callers may request an override with `allow_insecure_imports=true` if enabled by cluster administrator policy.

        **Retry semantics**: If a previous compilation attempt for the same suite and
        platform resulted in a `FAILED` artifact, uploading corrected sources via this
        endpoint automatically resets the artifact from `FAILED → PENDING` and prunes
        the stale Kubernetes Job before dispatching a fresh one. No manual database or
        cluster intervention is required — just re-upload and the control plane handles
        the rest.
      operationId: uploadAndBuild
      tags:
        - Builds
      parameters:
        - name: id
          in: path
          required: true
          description: Test Suite ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - source
              properties:
                source:
                  type: string
                  format: binary
                  description: Test scenario archive (.tar.gz or .zip)
                platform:
                  type: string
                  description: Target architecture (linux/amd64, linux/arm64, or all)
                  default: linux/amd64
                allow_insecure_imports:
                  type: boolean
                  description: Optional override to allow suspicious packages (e.g., syscall, unsafe) if permitted by cluster deployment policy
                  default: false
      responses:
        '202':
          description: Compilation build job accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BuildTriggerResponse'
        '400':
          description: Pre-build static analysis or validation failure (missing vuhive dependency, forbidden import, illegal package main, missing scenario contract, or invalid archive)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Import blocklist override forbidden by cluster policy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Test suite not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/suites/{id}/artifacts:
    get:
      summary: List compiled artifacts
      description: Returns all compiled binary artifacts associated with a test suite.
      operationId: listArtifacts
      tags:
        - Builds
      parameters:
        - name: id
          in: path
          required: true
          description: Test Suite ID
          schema:
            type: string
      responses:
        '200':
          description: List of compiled artifacts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArtifactListResponse'
  /api/v1/profiles:
    get:
      summary: List runner profiles
      description: Retrieves all registered runner scheduling and resource profiles.
      operationId: listProfiles
      tags:
        - Profiles
      responses:
        '200':
          description: List of runner profiles
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileListResponse'
    post:
      summary: Create runner profile
      description: Registers a new runner resource, affinity, and toleration profile.
      operationId: createProfile
      tags:
        - Profiles
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProfileRequest'
      responses:
        '201':
          description: Profile successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileResponse'
  /api/v1/profiles/{id}:
    get:
      summary: Get runner profile
      description: Retrieves details of a specific runner profile.
      operationId: getProfile
      tags:
        - Profiles
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Runner profile details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileResponse'
        '404':
          description: Profile not found
    put:
      summary: Update runner profile
      description: Updates resource constraints and affinity rules of an existing profile.
      operationId: updateProfile
      tags:
        - Profiles
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
      responses:
        '200':
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileResponse'
    delete:
      summary: Delete runner profile
      operationId: deleteProfile
      tags:
        - Profiles
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Profile deleted
  /api/v1/schedules:
    get:
      summary: List active test schedules
      description: Returns all active schedules and their associated Kubernetes CronJobs.
      operationId: listSchedules
      tags:
        - Schedules
      responses:
        '200':
          description: List of active schedules
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScheduleListResponse'
    post:
      summary: Create scheduled test execution
      description: Registers a recurring schedule and manifests a native batch/v1 CronJob in Kubernetes.
      operationId: createSchedule
      tags:
        - Schedules
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateScheduleRequest'
      responses:
        '201':
          description: Schedule created and CronJob manifested
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScheduleResponse'
  /api/v1/schedules/{id}:
    get:
      summary: Get test schedule
      operationId: getSchedule
      tags:
        - Schedules
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Schedule details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScheduleResponse'
    put:
      summary: Update schedule cadence
      operationId: updateSchedule
      tags:
        - Schedules
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateScheduleRequest'
      responses:
        '200':
          description: Schedule updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScheduleResponse'
    delete:
      summary: Delete schedule
      description: Deletes schedule record and terminates underlying Kubernetes CronJob.
      operationId: deleteSchedule
      tags:
        - Schedules
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Schedule deleted
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: OIDC JWT access token issued by Keycloak. Pass in Authorization header as 'Bearer <token>'.
    OAuth2AuthorizationCode:
      type: oauth2
      description: Keycloak Authorization Code flow with PKCE
      flows:
        authorizationCode:
          authorizationUrl: http://localhost:8080/realms/vuhive/protocol/openid-connect/auth
          tokenUrl: http://localhost:8080/realms/vuhive/protocol/openid-connect/token
          scopes:
            openid: OpenID Connect identity scope
            profile: User profile and username
            email: User email address
    OAuth2ClientCredentials:
      type: oauth2
      description: Keycloak M2M Client Credentials grant for automated runners and background services
      flows:
        clientCredentials:
          tokenUrl: http://localhost:8080/realms/vuhive/protocol/openid-connect/token
          scopes:
            openid: OpenID Connect identity scope
  schemas:
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          example: "resource not found"
...
    PresignedURLResponse:
      type: object
      required:
        - download_url
        - expires_in_seconds
      properties:
        download_url:
          type: string
          example: "https://minio.vuhive.local/artifacts/runs/run-123/summary.json?X-Amz-Algorithm=..."
        expires_in_seconds:
          type: integer
          example: 900
...
    RunListResponse:
      type: object
      required:
        - runs
        - count
        - total
        - limit
        - offset
      properties:
        runs:
          type: array
          items:
            $ref: '#/components/schemas/RunResponse'
        count:
          type: integer
          example: 10
        total:
          type: integer
          format: int64
          example: 45
        limit:
          type: integer
          example: 10
        offset:
          type: integer
          example: 0
    AbortRunRequest:
      type: object
      properties:
        reason:
          type: string
          example: "high error rate detected"
        requested_by:
          type: string
          example: "operator-alice"
    TriggerRunRequest:
      type: object
      required:
        - suite_id
        - artifact_id
        - runner_profile_id
      properties:
        suite_id:
          type: string
          format: uuid
          description: Unique identifier of the target TestSuite
          example: "3fa85f64-5717-4562-b3fc-2c963f66afa6"
        artifact_id:
          type: string
          format: uuid
          description: Unique identifier of the compiled, READY Artifact
          example: "c7a6e118-20ab-48d6-953b-e01140026e61"
        configuration_id:
          type: string
          format: uuid
          description: Optional identifier of the attached scenario configuration
          example: "d1a85f64-5717-4562-b3fc-2c963f66afa7"
        runner_profile_id:
          type: string
          format: uuid
          description: Unique identifier of the RunnerProfile specifying execution resources
          example: "e8d665b1-2e67-4228-8ab6-79c5b248a31e"
...